move allowed pools to params -- let pool and pool_test focus on pool domain logic, not parameter & governance concerns
